Create `URL+ProjectDirectory` to get PROJECT_DIR at runtime
